+++ app/controllers/api/v8/apidocs_controller.rb
@@ -134,12 +134,12 @@ module Api
         Api::V8::Organizations::Courses::Users::SubmissionsController,
         Api::V8::SubmissionsController,
         Api::V8::UsersController,
-        Api::V8::Core::Organizations::CoursesController,
         Api::V8::Core::Organizations::Courses::UnlocksController,
         Api::V8::Core::ExercisesController,
         Api::V8::Core::Exercises::SolutionsController,
         Api::V8::Core::SubmissionsController,
         Api::V8::Core::ExercisesController,
+        Api::V8::Core::CoursesController,
         Course,
         Exercise,
         AvailablePoint,
